Revert "Trying to add link to top liker profile"
