--- frontend/src/scenes/forgotPassword/ForgotPassword.jsx
@@ -1,72 +1,0 @@
-import React, { useState } from 'react';
-import { Box, Button, TextField, Typography, useTheme } from "@mui/material";
-import { useNavigate } from "react-router-dom";
-
-const ForgotPassword = () => {
-  const [email, setEmail] = useState('');
-  const [message, setMessage] = useState('');
-  const { palette } = useTheme();
-  const navigate = useNavigate();
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    try {
-      const response = await fetch(`${process.env.REACT_APP_URL_BACKEND}/auth/request-password-reset`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ email }),
-      });
-
-      const data = await response.json();
-      setMessage(data.message);
-    } catch (error) {
-      setMessage('Error sending reset email');
-    }
-  };
-
-  return (
-    <Box>
-      <Typography variant="h4">Forgot Password</Typography>
-      <form onSubmit={handleSubmit}>
-        <TextField
-          label="Email"
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-          required
-          fullWidth
-        />
-        {message && <Typography>{message}</Typography>}
-        <Button
-          fullWidth
-          type="submit"
-          sx={{
-            m: "2rem 0",
-            p: "1rem",
-            backgroundColor: palette.primary.main,
-            color: palette.background.alt,
-            "&:hover": { color: palette.primary.main },
-          }}
-        >
-          Send Reset Link
-        </Button>
-        <Typography
-          onClick={() => navigate("/login")}
-          sx={{
-            textDecoration: "underline",
-            color: palette.primary.main,
-            "&:hover": {
-              cursor: "pointer",
-              color: palette.primary.light,
-            },
-          }}
-        >
-          Login here.
-        </Typography>
-      </form>
-    </Box>
-  );
-};
-
-export default ForgotPassword;
