+++ frontend/src/scenes/homePage/index.jsx
@@ -4,8 +4,7 @@ import Navbar from "scenes/navbar";
 import UserWidget from "scenes/widgets/UserWidget";
 import MyPostWidget from "scenes/widgets/MyPostWidget";
 import PostsWidget from "scenes/widgets/PostsWidget";
-import FollowersWidget from "scenes/widgets/FollowersWidget";
-import FollowingWidget from "scenes/widgets/FollowingWidget";
+import FriendListWidget from "scenes/widgets/FriendListWidget";
 
 const HomePage = () => {
   const isNonMobileScreens = useMediaQuery("(min-width:1000px)");
@@ -34,9 +33,7 @@ const HomePage = () => {
         {isNonMobileScreens && (
           <Box flexBasis="26%">
             <Box m="2rem 0" />
-            <FollowersWidget userId={_id} /> {/* Display Followers */}
-            <Box m="2rem 0" />
-            <FollowingWidget userId={_id} /> {/* Display Following */}
+            <FriendListWidget userId={_id} />
           </Box>
         )}
       </Box>
