+++ frontend/src/scenes/loginPage/Form.jsx
@@ -1,0 +1,306 @@
+import { useState } from "react";
+import { Box, Button, TextField, Typography, useTheme } from "@mui/material";
+import { Formik } from "formik";
+import * as yup from "yup";
+import { useNavigate } from "react-router-dom";
+import { useDispatch } from "react-redux";
+import { setLogin } from "state";
+import FlexBetween from "components/FlexBetween";
+
+const registerSchema = yup.object().shape({
+  firstName: yup.string(),
+  lastName: yup.string(),
+  email: yup.string().email("Invalid email").required("Required"),
+  password: yup.string().min(8, "At least 8 characters are required").required("Required"),
+  location: yup.string(),
+  username: yup.string().required("Required"),
+  dateOfBirth: yup.date(),
+});
+
+const loginSchema = yup.object().shape({
+  identifier: yup.string().required("Required"),
+  password: yup.string().min(8, "At least 8 characters are required").required("Required"),
+});
+
+const initialValuesRegister = {
+  firstName: "",
+  lastName: "",
+  email: "",
+  password: "",
+  location: "",
+  username: "",
+  dateOfBirth: "",
+};
+
+const initialValuesLogin = {
+  identifier: "",
+  password: "",
+};
+
+const Form = () => {
+  const [pageType, setPageType] = useState("login");
+  const { palette } = useTheme();
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
+  const isLogin = pageType === "login";
+  const isRegister = pageType === "register";
+
+  const register = async (values, onSubmitProps) => {
+    const formData = {
+      firstName: values.firstName,
+      lastName: values.lastName,
+      email: values.email,
+      password: values.password,
+      location: values.location,
+      username: values.username,
+      dateOfBirth: values.dateOfBirth,
+    };
+
+    try {
+      const savedUserResponse = await fetch(
+        `${process.env.REACT_APP_URL_BACKEND}/auth/register`,
+        {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify(formData),
+        }
+      );
+      const savedUser = await savedUserResponse.json();
+      onSubmitProps.resetForm();
+
+      if (savedUser) {
+        setPageType("login");
+      }
+    } catch (error) {
+      console.error("Registration error:", error);
+      // Handle registration error
+    }
+  };
+
+  
+  const login = async (values, onSubmitProps) => {
+    const { identifier, password } = values;
+
+    try {
+      const loggedInResponse = await fetch(`${process.env.REACT_APP_URL_BACKEND}/auth/login`, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({
+          identifier: identifier,
+          password: password,
+        }),
+      });
+
+      if (!loggedInResponse.ok) {
+        throw new Error("Username or password incorrect");
+      }
+
+      const loggedIn = await loggedInResponse.json();
+      onSubmitProps.resetForm();
+
+      if (loggedIn) {
+        dispatch(
+          setLogin({
+            user: loggedIn.user,
+            token: loggedIn.token,
+          })
+        );
+        navigate("/home");
+      } else {
+        throw new Error("Login failed");
+      }
+    } catch (error) {
+      console.error("Login error:", error.message);
+      alert(error.message);
+    }
+  };
+
+  const handleFormSubmit = async (values, onSubmitProps) => {
+    if (isLogin) await login(values, onSubmitProps);
+    if (isRegister) await register(values, onSubmitProps);
+  };
+
+  return (
+    <Formik
+      onSubmit={handleFormSubmit}
+      initialValues={isLogin ? initialValuesLogin : initialValuesRegister}
+      validationSchema={isLogin ? loginSchema : registerSchema}
+    >
+      {({
+        values,
+        errors,
+        touched,
+        handleBlur,
+        handleChange,
+        handleSubmit,
+        setFieldValue,
+        resetForm,
+      }) => (
+        <form onSubmit={handleSubmit}>
+          <Box
+            display="grid"
+            gap="30px"
+            gridTemplateColumns="repeat(4, minmax(0, 1fr))"
+          >
+            {isRegister && (
+              <>
+                <TextField
+                  label="First Name"
+                  onBlur={handleBlur}
+                  onChange={handleChange}
+                  value={values.firstName}
+                  name="firstName"
+                  error={touched.firstName && errors.firstName}
+                  helperText={touched.firstName && errors.firstName}
+                  sx={{ gridColumn: "span 2" }}
+                />
+                <TextField
+                  label="Last Name"
+                  onBlur={handleBlur}
+                  onChange={handleChange}
+                  value={values.lastName}
+                  name="lastName"
+                  error={touched.lastName && errors.lastName}
+                  helperText={touched.lastName && errors.lastName}
+                  sx={{ gridColumn: "span 2" }}
+                />
+                <TextField
+                  label="Location"
+                  onBlur={handleBlur}
+                  onChange={handleChange}
+                  value={values.location}
+                  name="location"
+                  error={touched.location && errors.location}
+                  helperText={touched.location && errors.location}
+                  sx={{ gridColumn: "span 4" }}
+                />
+                <TextField
+                  label="Username *"
+                  onBlur={handleBlur}
+                  onChange={handleChange}
+                  value={values.username}
+                  name="username"
+                  error={touched.username && errors.username}
+                  helperText={touched.username && errors.username}
+                  sx={{ gridColumn: "span 4" }}
+                />
+                <TextField
+                  label="Date of Birth"
+                  type="date"
+                  InputLabelProps={{ shrink: true }}
+                  onBlur={handleBlur}
+                  onChange={handleChange}
+                  value={values.dateOfBirth}
+                  name="dateOfBirth"
+                  error={touched.dateOfBirth && errors.dateOfBirth}
+                  helperText={touched.dateOfBirth && errors.dateOfBirth}
+                  sx={{ gridColumn: "span 4" }}
+                />
+                <TextField
+                  label="Email *"
+                  onBlur={handleBlur}
+                  onChange={handleChange}
+                  value={values.email}
+                  name="email"
+                  error={touched.email && errors.email}
+                  helperText={touched.email && errors.email}
+                  sx={{ gridColumn: "span 4" }}
+                />
+                <TextField
+                  label="Password *"
+                  type="password"
+                  onBlur={handleBlur}
+                  onChange={handleChange}
+                  value={values.password}
+                  name="password"
+                  error={touched.password && errors.password}
+                  helperText={touched.password && errors.password}
+                  sx={{ gridColumn: "span 4" }}
+                />
+              </>
+            )}
+            {isLogin && (
+              <>
+                <TextField
+                  label="Email or Username"
+                  onBlur={handleBlur}
+                  onChange={handleChange}
+                  value={values.identifier}
+                  name="identifier"
+                  error={touched.identifier && Boolean(errors.identifier)}
+                  helperText={touched.identifier && errors.identifier}
+                  sx={{ gridColumn: "span 4" }}
+                />
+                <TextField
+                  label="Password"
+                  type="password"
+                  onBlur={handleBlur}
+                  onChange={handleChange}
+                  value={values.password}
+                  name="password"
+                  error={touched.password && errors.password}
+                  helperText={touched.password && errors.password}
+                  sx={{ gridColumn: "span 4" }}
+                />
+                <Typography
+                  onClick={() => navigate("/forgot-password")}
+                  sx={{
+                    textDecoration: "underline",
+                    color: palette.primary.main,
+                    "&:hover": {
+                      cursor: "pointer",
+                      color: palette.primary.light,
+                    },
+                    gridColumn: "span 4",
+                    textAlign: "right",
+                  }}
+                >
+                  Forgot Password?
+                </Typography>
+              </>
+            )}
+          </Box>
+
+          {/* BUTTONS */}
+          <Box>
+            <Button
+              fullWidth
+              type="submit"
+              sx={{
+                m: "2rem 0",
+                p: "1rem",
+                backgroundColor: palette.primary.main,
+                color: palette.background.alt,
+                "&:hover": { color: palette.primary.main },
+              }}
+            >
+              {isLogin ? "LOGIN" : "REGISTER"}
+            </Button>
+            <Typography
+              onClick={() => {
+                setPageType(isLogin ? "register" : "login");
+                resetForm();
+              }}
+              sx={{
+                textDecoration: "underline",
+                color: palette.primary.main,
+                "&:hover": {
+                  cursor: "pointer",
+                  color: palette.primary.light,
+                },
+              }}
+            >
+              {isLogin
+                ? "Don't have an account? Sign Up here."
+                : "Already have an account? Login here."}
+            </Typography>
+          </Box>
+        </form>
+      )}
+    </Formik>
+  );
+};
+
+export default Form;
