+++ frontend/src/scenes/loginPage/TiktourLogo.png
