--- frontend/src/scenes/resetPassword/ResetPassword.jsx
@@ -1,76 +1,0 @@
-import React, { useState } from 'react';
-import { Box, Button, TextField, Typography, useTheme } from "@mui/material";
-import { useSearchParams, useNavigate } from "react-router-dom";
-
-const ResetPassword = () => {
-  const [newPassword, setNewPassword] = useState('');
-  const [message, setMessage] = useState('');
-  const [searchParams] = useSearchParams();
-  const { palette } = useTheme();
-  const navigate = useNavigate();
-
-  const token = searchParams.get('token');
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    try {
-      const response = await fetch(`${process.env.REACT_APP_URL_BACKEND}/auth/reset-password`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ token, newPassword }),
-      });
-
-      const data = await response.json();
-      setMessage(data.message);
-    } catch (error) {
-      setMessage('Error resetting password');
-    }
-  };
-
-  return (
-    <Box>
-      <Typography variant="h4">Reset Password</Typography>
-      <form onSubmit={handleSubmit}>
-        <TextField
-          label="New Password"
-          type="password"
-          value={newPassword}
-          onChange={(e) => setNewPassword(e.target.value)}
-          required
-          fullWidth
-        />
-        {message && <Typography>{message}</Typography>}
-        <Button
-          fullWidth
-          type="submit"
-          sx={{
-            m: "2rem 0",
-            p: "1rem",
-            backgroundColor: palette.primary.main,
-            color: palette.background.alt,
-            "&:hover": { color: palette.primary.main },
-          }}
-        >
-          Reset Password
-        </Button>
-        <Typography
-          onClick={() => navigate("/login")}
-          sx={{
-            textDecoration: "underline",
-            color: palette.primary.main,
-            "&:hover": {
-              cursor: "pointer",
-              color: palette.primary.light,
-            },
-          }}
-        >
-          Login here.
-        </Typography>
-      </form>
-    </Box>
-  );
-};
-
-export default ResetPassword;
