+++ frontend/src/scenes/widgets/MyPostWidget.jsx
@@ -1,4 +1,3 @@
-import React, { useState } from 'react';
 import {
   EditOutlined,
   DeleteOutlined,
@@ -7,7 +6,6 @@ import {
   ImageOutlined,
   MicOutlined,
   MoreHorizOutlined,
-  LocationOnOutlined
 } from "@mui/icons-material";
 import {
   Box,
@@ -23,18 +21,15 @@ import FlexBetween from "components/FlexBetween";
 import Dropzone from "react-dropzone";
 import UserImage from "components/UserImage";
 import WidgetWrapper from "components/WidgetWrapper";
-import LocationAutocomplete from "./LocationAutocomplete";
+import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { setPosts } from "state";
 
 const MyPostWidget = ({ picturePath }) => {
   const dispatch = useDispatch();
   const [isImage, setIsImage] = useState(false);
-  let [location, setLocation] = useState(""); // Location state
-  const [addedLocation, setAddedLocation] = useState(false);
   const [image, setImage] = useState(null);
   const [post, setPost] = useState("");
-  const [showLocationAutocomplete, setShowLocationAutocomplete] = useState(false); // To toggle autocomplete display
   const { palette } = useTheme();
   const { _id } = useSelector((state) => state.user);
   const token = useSelector((state) => state.token);
@@ -46,8 +41,6 @@ const MyPostWidget = ({ picturePath }) => {
     const formData = new FormData();
     formData.append("userId", _id);
     formData.append("description", post);
-    formData.append("location", location); // Add location to formData
-
     if (image) {
       formData.append("picture", image);
       formData.append("picturePath", image.name);
@@ -62,39 +55,23 @@ const MyPostWidget = ({ picturePath }) => {
     dispatch(setPosts({ posts }));
     setImage(null);
     setPost("");
-    setLocation(""); // Reset location after posting
-
-    alert("The post was posted successfully!");
   };
 
   return (
     <WidgetWrapper>
       <FlexBetween gap="1.5rem">
         <UserImage image={picturePath} />
-        <FlexBetween
+        <InputBase
+          placeholder="What's on your mind..."
+          onChange={(e) => setPost(e.target.value)}
+          value={post}
           sx={{
-            flexDirection: 'column', // Arrange children in a column
-            alignItems: 'flex-start', // Align children to start of the cross-axis (horizontal alignment)
-            width: '100%' // Ensure the container takes full width
+            width: "100%",
+            backgroundColor: palette.neutral.light,
+            borderRadius: "2rem",
+            padding: "1rem 2rem",
           }}
-        >
-          <b style={ {color: mediumMain}}>       
-            <span style={{ color: 'red' }}>* </span>
-            Description
-          </b>
-          <InputBase
-            placeholder="What's on your mind..."
-            onChange={(e) => setPost(e.target.value)}
-            value={post}
-            sx={{
-              width: "100%", // Full width of the container
-              backgroundColor: palette.neutral.light,
-              borderRadius: "2rem",
-              padding: "1rem 2rem",
-              marginTop: "0.5rem" // Add some spacing between the label and the input
-            }}
-          />
-    </FlexBetween>
+        />
       </FlexBetween>
       {isImage && (
         <Box
@@ -140,19 +117,6 @@ const MyPostWidget = ({ picturePath }) => {
           </Dropzone>
         </Box>
       )}
-      <Box>
-        {addedLocation && <p>{location}</p>}
-        {showLocationAutocomplete && (
-          <LocationAutocomplete
-            onSelectLocation={(selectedLocation) => {
-              setLocation(selectedLocation);
-              setShowLocationAutocomplete(false); // Hide autocomplete after selection
-              location = selectedLocation;
-              setAddedLocation(true);
-            }}
-          />
-        )}
-      </Box>
 
       <Divider sx={{ margin: "1.25rem 0" }} />
 
@@ -167,24 +131,31 @@ const MyPostWidget = ({ picturePath }) => {
           </Typography>
         </FlexBetween>
 
-        <FlexBetween gap="0.25rem" onClick={() => {
-          setShowLocationAutocomplete(!showLocationAutocomplete);
-        }}>
-          <LocationOnOutlined sx={{ color: mediumMain }} />
-          <Typography
-            color={mediumMain}
-            sx={{ "&:hover": { cursor: "pointer", color: medium } }}
-          >
-            <b>       
-            <span style={{ color: 'red' }}>* </span>
-            Location
-          </b>
-          </Typography>
-        </FlexBetween>
+        {isNonMobileScreens ? (
+          <>
+            <FlexBetween gap="0.25rem">
+              <GifBoxOutlined sx={{ color: mediumMain }} />
+              <Typography color={mediumMain}>Clip</Typography>
+            </FlexBetween>
 
+            <FlexBetween gap="0.25rem">
+              <AttachFileOutlined sx={{ color: mediumMain }} />
+              <Typography color={mediumMain}>Attachment</Typography>
+            </FlexBetween>
+
+            <FlexBetween gap="0.25rem">
+              <MicOutlined sx={{ color: mediumMain }} />
+              <Typography color={mediumMain}>Audio</Typography>
+            </FlexBetween>
+          </>
+        ) : (
+          <FlexBetween gap="0.25rem">
+            <MoreHorizOutlined sx={{ color: mediumMain }} />
+          </FlexBetween>
+        )}
 
         <Button
-          disabled={!post || !addedLocation}
+          disabled={!post}
           onClick={handlePost}
           sx={{
             color: palette.background.alt,
