+++ frontend/src/scenes/widgets/UserWidget.jsx
@@ -11,7 +11,6 @@ import WidgetWrapper from "components/WidgetWrapper";
 import { useSelector } from "react-redux";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { date } from "yup";
 
 const UserWidget = ({ userId, picturePath }) => {
   const [user, setUser] = useState(null);
@@ -33,7 +32,7 @@ const UserWidget = ({ userId, picturePath }) => {
 
   useEffect(() => {
     getUser();
-  }, [userId, token]);
+  }, []); 
 
   if (!user) {
     return null;
@@ -46,10 +45,8 @@ const UserWidget = ({ userId, picturePath }) => {
     rank,
     username,
     dateOfBirth,
+    friends,
   } = user;
-  
-  const fixedBirthdayDate = dateOfBirth === null ? "No birthday given" : dateOfBirth.slice(0,10);
-
 
   return (
     <WidgetWrapper>
@@ -73,8 +70,9 @@ const UserWidget = ({ userId, picturePath }) => {
                 },
               }}
             >
-              {firstName=== null ? "firstName" : firstName} {lastName === null ? "lastName" : lastName}
+              {firstName} {lastName}
             </Typography>
+            <Typography color={medium}>{friends.length} friends</Typography>
           </Box>
         </FlexBetween>
         <ManageAccountsOutlined />
@@ -99,20 +97,22 @@ const UserWidget = ({ userId, picturePath }) => {
       {/* THIRD ROW */}
       <Box p="1rem 0">
         <FlexBetween mb="0.5rem">
-          <Typography color={medium}>Birthday Date</Typography>
+          <Typography color={medium}>birthday date</Typography>
           <Typography color={main} fontWeight="500">
-            {fixedBirthdayDate}
+            {dateOfBirth}
           </Typography>
         </FlexBetween>
         <FlexBetween>
-          <Typography color={medium}>Username</Typography>
+          <Typography color={medium}>User Name</Typography>
           <Typography color={main} fontWeight="500">
             {username}
           </Typography>
         </FlexBetween>
       </Box>
 
-      {/* SOCIAL PROFILES */}
+      <Divider />
+
+      {/* FOURTH ROW */}
       <Box p="1rem 0">
         <Typography fontSize="1rem" color={main} fontWeight="500" mb="1rem">
           Social Profiles
